+++ federation/mod.ts
@@ -15,3 +15,4 @@ export * from "./kv.ts";
 export * from "./middleware.ts";
 export * from "./mq.ts";
 export * from "./router.ts";
+export { type SenderKeyPair } from "./send.ts";