+++ federation/queue.ts
@@ -1,7 +1,11 @@
-export interface OutboxMessage {
-  type: "outbox";
+export interface SenderKeyJwkPair {
   keyId: string;
   privateKey: JsonWebKey;
+}
+
+export interface OutboxMessage {
+  type: "outbox";
+  keys: SenderKeyJwkPair[];
   activity: unknown;
   inbox: string;
   trial: number;